PixelParts: Add preference fragment for changing saturation level
To apply the saturation changes, the method executes the command
"service call SurfaceFlinger 1022 f [saturation]" using
Runtime.getRuntime().exec() when the seekbar progress is set.
However, when the devices color mode is set to 'saturated',
a saturation value of 1.0 does not register any changes. To
overcome this, a value of 1.001 is set for a seek bar value of 100.

Co-authored-by: SKULSHADY <anushekprasal@gmail.com>
Signed-off-by: AnierinB <anierin@evolution-x.org>