PixelParts: Display the current lux level in AutoHBM fragment
During onSensorChanged, the lux level is set as the summary for the
current_lux_level preference.

Listener is registered when enabling the main switch and when the
fragment is resumed (only if the main switch is on). Listener is
unregistered when the main switch is turned off and when the fragment
is paused (only if the main switch is on).

Signed-off-by: AnierinB <anierin@evolution-x.org>